[refs #85843] Re-organize sync functions
