+++ fcs/sync/__init__.py
@@ -11,4 +11,4 @@ class InvalidResponse(Exception):
     pass
 
 
-import fcs.sync.fgases
+import fcs.sync.commands
