--- fcs/sync/utils.py
@@ -1,7 +1,0 @@
-def update_obj(obj, d):
-    if not d:
-        obj = None
-    else:
-        for name, value in d.iteritems():
-            setattr(obj, name, value)
-    return obj
